remove the tsa property from the JarSigner calls and move it to the Abstract one to be able to use it there. It doesn't fully belong there but then we need to split up processArchive and and do the actual call in the sub class again if the subclass wants to override this.
added retry and delay also to it.

Now it splits the tsa url and if there are more then one it will go to
the next if the current tsa url fails. if all of them fails it will wait
for the delay and then retry it again the number of times the retry is
set. (defaults to 1)